hal: create Effect worker thread only if needed
Offload effects and effects operating with bypass
flags don't operate on the data, as effect will be
actually applied in DSP side, for such scenarios
there is no need to init the FMQ and start worker
threads for data processing. Worker thread will be
only created where data processing is needed.

Remove the statusDepth parameter from EffectContext
creation as it is always 1.

Improve the logging as well, change log types based
on need.

Change-Id: I00cf79e42e3c514588487755360899f17349f8ff