ahal core: Fix handling of external devices disconnection
A mix port can be patched to multiple connected device ports. Thus, when
disconnecting an external device and removing the connected port, the
profiles of the mix port can only be cleared iff there are no more
connected device ports patched to it, and it did not have profiles prior to
connection of the first device.

Also, make connection/disconnection logs more appealing for debugging

This is a manual cherry-pick of 0e128dd3fea79d6ec4549abc44b3ee0777aebced
of reference Audio HAL.

Change-Id: I9ec66c94a4407a9c43e0f66dbc6525d9f6cb1ec4