core: fix the StreamWorker exit when client doesn't close
client is always expected to close the StreamWorker before
Stream(In|Out) destruction. But, Service shouldn't rely on this.

During the destruction of Stream(In|Out), if the StreamWorker
is not closed by the client, the HAL tries to exit the
StreamWorker's I/O thread by sending the halReservedExit command
to the command FMQ where I/O thread is waiting for commands.

While handling the halReservedExit, shutdown API Implementation
already got freed. Hence, add the default Implementation for
shutdown API.

FYI, the resources are released by Stream(In|Out)Primary's
destructor which happened before destruction of Stream(In|Out).

Change-Id: I20b4e55e3ec4187827602f04a16a5a5cb60d3df3