Hal: core: change kFMQMinFrameSize to 160 instead of 256
8k recording has distortion because DSP only returns 480
frames data though we are requesting 512 frames. because
DSP used multiples of 32 after calculating by sample rate
and format and channel transferred from hal as buffer size.
So the kFMQMinFrameSize limit of 256 brings recording distortion.

Need to consider the framesize calculation of minimum sample rate 8k.
Change kFMQMinFrameSize to 160 instead of 256 to fix it.

Change-Id: I4756565e60cb44c5529db7c7d0d6834e89e5f388