core: add latency support
Framework expects the latency of the stream
via StreamDescriptor::Reply::latencyMs.

This latencyMs, understood as, the hardware delay
nothing but HLOS HAL delay + DSP pipeline delay +
codec delay.

Generally, HLOS HAL delay and codec delay are small
and ideally constant. But, they are not 0.

This delay is majorly contributed by DSP pipeline
modules.

In most of the usecases, the DSP (Graph) pipeline delay
is precaluclated.

Change-Id: Id2cff80c74c2d5d9a398385fc6676eeb1270d6bd