hal: Fix issue w.r.t draining state and mmap issues
In current scenarios, for non offload
streams once we are in draining state for
any subsequent commands we move state to
Idle and those commands can fail in Idle
state.

For non-offload streams keep state in draining
itself, because as part of stream standby we
get pause/flush/standby commands which moves
stream state to IDLE. This aligns with stream
state machine.

For mmap, drain itself is stop command

Change-Id: I4b2d50d6993d2344ac7f8dd45794f305343dcded