hal: Fix issues w.r.t stream error state
currently if any command like pause/flush/drain
fails, we update error and also set state to error
And subsequent burst commands are ignored in burst
state causing data loss

As framework is more interested in succcess/failure
of commands, just send the error to framework and
dont set the streams state to error as its not really
fatal.

This ensures that HAL is not treating any command failure
as hard error.

Hard error is any case that stream is not recoverable, no
further actions performed with the stream. The only further
action is stream destruction.

Change-Id: Ic8f06287e177ddd32a7e211f3eaf760b603a7f68