AHAL: fix state machine for drain command
Only set stream state machine to IDLE when receiving the drain
ready callback with DRAIN_ALL type.

If the callback of drain type DRAIN_EARLY_NOTIFY coming, that
means all the data is about to be consumed, but it has not been
consumed yet. So, should not set the stream state to IDLE, because
the user side may send pause, flush, resume command during this
time.

Change-Id: I6714efb1760f796572fe2e78fa26194fddf57430