hal: core: always send routing for BT devices
For streams active on BT devices, there might be internal routing
done in PAL to dummy devices. If APM sends routing for the stream
to BT device, it gets rejected in AHAL due to earlier routing also
being on the BT device. This will keep the stream on dummy device
which is not desirable.

Fixed this by always allowing routing to BT devices and let PAL
decide on whether to do device switch or not.

Change-Id: If487901056fd6e98799ac6d9c2569e9959f8844f