hal: core: always send routing for BT devices including SCO device
For streams active on BT devices, there might be internal routing
done in PAL to dummy devices. If APM sends routing for the stream
to BT device, it gets rejected in AHAL due to earlier routing also
being on the BT device. This will keep the stream on dummy device
which is not desirable.

Fixed this by always allowing routing to BT devices including SCO
device and let PAL decide on whether to do device switch or not.

Change-Id: I3c8a06a4d8f1f37adb026703474055879e1461e6